split partitioner and strategy out
