+++ types/assignment/partitioners/hash_crc32.d.ts
@@ -1,0 +1,16 @@
+
+
+declare module "assignment/partitioners/hash_crc32" {
+    export class HashCRC32Partitioner {
+        constructor();
+        /**
+         * The partition method receives 3 arguments: 
+         * * the topic name, 
+         * * an array with topic partitions, and 
+         * * the message (useful to partition by key, etc.). 
+         * The call to partition can be sync or async (return a Promise).
+         */
+        partition(topic: string, partitions: number[], message: string): Promise<number>;
+
+    }
+}
