+++ types/assignment/strategies/default.d.ts
@@ -1,0 +1,17 @@
+
+
+declare module "assignment/strategies/default" {
+    export class AbstractAssignmentStrategy {
+        constructor();
+    }
+    /**
+     * simple round robin assignment strategy (default).
+     * 
+     * @export
+     * @class DefaultAssignmentStrategy
+     */
+    export class DefaultAssignmentStrategy extends AbstractAssignmentStrategy {
+        constructor();
+    }
+   
+}
