+++ types/assignment/strategies/weighted_round_robin.d.ts
@@ -1,0 +1,12 @@
+
+
+declare module "assignment/strategies/weighted_round_robin" {
+
+    /** 
+    * WeightedRoundRobinAssignmentStrategy weighted round robin assignment (based on wrr-pool).
+    */
+    export class WeightedRoundRobinAssignmentStrategy extends AbstractAssignmentStrategy {
+        constructor();
+    }
+
+}
