+++ types/index.d.ts
@@ -1,6 +1,7 @@
 /// <reference path="producer.d.ts" />
 /// <reference path="simple_consumer.d.ts" />
 /// <reference path="group_consumer.d.ts" />
+/// <reference path="group_admin.d.ts" />
 
 /// <reference path="kafka.d.ts" />
 /// <reference path="client.d.ts" />
@@ -8,17 +9,17 @@
 declare module "no-kafka" {
 
     export * from "kafka";
-    export { Producer } from "producer";
+    export { Producer, Result } from "producer";
     export { SimpleConsumer } from "simple_consumer";
     export { GroupConsumer } from "group_consumer";
 
-    // exports.GroupAdmin = require('./group_admin');
+    export { GroupAdmin } from "group_admin";
 
-    // exports.DefaultPartitioner = require('./assignment/partitioners/default');
-    // exports.HashCRC32Partitioner = require('./assignment/partitioners/hash_crc32');
+    export { DefaultPartitioner } from "assignment/partitioners/default";
+    export { HashCRC32Partitioner } from "assignment/partitioners/hash_crc32";
 
-    // exports.DefaultAssignmentStrategy = require('./assignment/strategies/default');
-    // exports.ConsistentAssignmentStrategy = require('./assignment/strategies/consistent');
-    // exports.WeightedRoundRobinAssignmentStrategy = require('./assignment/strategies/weighted_round_robin');
+    export { DefaultAssignmentStrategy } from "assignment/strategies/default";
+    export { ConsistentAssignmentStrategy } from "assignment/strategies/consistent";
+    export { WeightedRoundRobinAssignmentStrategy } from "assignment/strategies/weighted_round_robin";
 
 }
